feat: 시간 변환하는 함수 분리
